backend/requirements.txt
fastapi
fastapi[standard]
uvicorn[standard]
pydantic
streamlit
requests
python-dotenv
pandas
matplotlib
seaborn
langchain
langchain-core
langchain-community
langchain-ollama
faiss-cpu
langchain-groq